restruct: moving away from single init.lua
