+++ lua/config/init.lua
@@ -1,0 +1,5 @@
+return {
+    options = require 'config.options',
+    keymaps = require 'config.keymaps',
+    autocommands = require 'config.autocommands',
+}
