+++ lua/config/options.lua
@@ -1,0 +1,38 @@
+vim.g.mapleader = ' '
+vim.g.maplocalleader = ' '
+vim.g.have_nerd_font = true
+vim.o.number = true
+vim.o.relativenumber = true
+vim.o.mouse = 'a'
+vim.o.showmode = false
+vim.o.breakindent = true
+vim.o.undofile = true
+vim.o.ignorecase = true
+vim.o.smartcase = true
+vim.o.signcolumn = 'yes'
+vim.o.updatetime = 250
+vim.o.timeoutlen = 300
+vim.o.splitright = true
+vim.o.splitbelow = true
+vim.o.inccommand = 'split'
+vim.o.cursorline = true
+vim.o.scrolloff = 10
+vim.o.confirm = true
+vim.o.textwidth = 120
+
+-- Set TAB defaults
+vim.o.tabstop = 4 -- Number of spaces a <Tab> character shows as
+vim.o.shiftwidth = 4 -- Number of spaces used for autoindent
+vim.o.softtabstop = 4 -- Number of spaces a <Tab> keypress inserts
+vim.o.expandtab = true -- Use spaces instead of tabs
+vim.o.autoindent = true
+vim.o.smartindent = true
+
+-- Sets how neovim will display certain whitespace characters in the editor.
+vim.o.list = true
+vim.opt.listchars = { tab = '» ', trail = '·', nbsp = '␣' }
+
+-- Sync clipboard between OS and Neovim.
+vim.schedule(function()
+    vim.o.clipboard = 'unnamedplus'
+end)
