--- lua/custom/plugins/init.lua
@@ -1,9 +1,0 @@
--- init.lua
---
--- This file just organizes my custom plugins into seperate files. Might put stuff here later idk.
-
-return {
-    require 'custom.plugins.colorscheme',
-    require 'custom.plugins.ui',
-    require 'custom.plugins.code',
-}
