+++ lua/plugins/lualine.lua
@@ -1,0 +1,10 @@
+return {
+    {
+        'nvim-lualine/lualine.nvim',
+        dependencies = { 'nvim-tree/nvim-web-devicons' },
+        opts = {
+            use_icons = true,
+            theme = 'gruber-darker',
+        },
+    },
+}
