+++ lua/plugins/vimtex.lua
@@ -1,0 +1,12 @@
+return {
+    -- VIMTEX
+    {
+        'lervag/vimtex',
+        lazy = false, -- we don't want to lazy load VimTeX
+        -- tag = "v2.15", -- uncomment to pin to a specific release
+        init = function()
+            -- VimTeX configuration goes here, e.g.
+            vim.g.vimtex_view_method = 'mupdf'
+        end,
+    },
+}
