+++ plugins.txt
@@ -1,0 +1,39 @@
+
+mason-org/mason-lspconfig.nvim
+WhoIsSethDaniel/mason-tool-installer.nvim
+nvim-lualine/lualine.nvim
+nvim-treesitter/nvim-treesitter
+lervag/vimtex
+stevearc/conform.nvim
+Jezda1337/nvim-html-css
+L3MON4D3/LuaSnip
+hrsh7th/nvim-cmp
+folke/todo-comments.nvim
+MeanderingProgrammer/render-markdown.nvim
+nvim-telescope/telescope.nvim
+mfussenegger/nvim-dap
+nvim-lua/plenary.nvim
+rcarriga/nvim-dap-ui
+pathlib.nvim
+jay-babu/mason-nvim-dap.nvim
+nvim-neorg/neorg
+leoluz/nvim-dap-go
+MunifTanjim/nui.nvim
+lukas-reineke/indent-blankline.nvim
+nvim-neotest/nvim-nio
+nvim-neorg/lua-utils.nvim
+windwp/nvim-autopairs
+lewis6991/gitsigns.nvim
+folke/which-key.nvim
+nvim-telescope/telescope-fzf-native.nvim
+nvim-telescope/telescope-ui-select.nvim
+folke/lazy.nvim
+saghen/blink.cmp
+blazkowolf/gruber-darker.nvim
+folke/lazydev.nvim
+j-hui/fidget.nvim
+nvim-tree/nvim-web-devicons
+neovim/nvim-lspconfig
+echasnovski/mini.nvim
+mfussenegger/nvim-lint
+mason-org/mason.nvim
