+++ util/list-plugins.lua
@@ -1,0 +1,3 @@
+for _, plugin in ipairs(require('lazy').plugins()) do
+    vim.cmd('redir >> plugins.txt | echo "' .. plugin[1] .. '" | redir END')
+end
